Extract DOM beautify logic into a mixin that the DOMDBUtils and CPODOMDBUtils can share
